Renamed configure_controller plan to configure_driver
