+++ tests/devices/unit_tests/electron_analyser/test_base_analayser_io.py
@@ -13,7 +13,9 @@ from dodal.devices.electron_analyser.specs_analyser_io import (
     SpecsAnalyserDriverIO,
 )
 from dodal.devices.electron_analyser.specs_region import SpecsSequence
-from dodal.devices.electron_analyser.util import to_kinetic_energy
+from dodal.devices.electron_analyser.util import (
+    to_kinetic_energy,
+)
 from dodal.devices.electron_analyser.vgscienta_analyser_io import (
     VGScientaAnalyserDriverIO,
 )
@@ -54,6 +56,7 @@ def analyser_type(
 
 @pytest.mark.parametrize("region", TEST_SEQUENCE_REGION_NAMES, indirect=["region"])
 def test_analyser_to_kinetic_energy(
+    sim_analyser_driver: AbstractAnalyserDriverIO,
     region: AbstractBaseRegion,
     excitation_energy: float,
 ) -> None:
